// src/utils/types.ts
export type ColorQuantity = { color: string; quantity: number }
export type Variant = { size: string; colors: ColorQuantity[] }

export type Product = {
  id: string
  name: string
  description: string
  price: number
  category: string
  material: string
  brand: string
  stock: number
  variants: Variant[]
  images: string[]
  vendorid: string
  createdat: string
  rating: number
  vendor: string
  updated_at: string
}
// ...
export type CartItemType = {
  images: string[]
  name: string
  price: number
  color: string
  size: string
  availableVariants: Variant[]
  amount: number
  id: string
  vendor: string
  vendorid: string
  category: string
}